requirements.txt
pillow~=11.1.0
platformdirs~=4.3.6
chardet~=5.2.0
easyocr~=1.7.2
numpy~=2.2.2